+++ core/datetime/src/commonMain/kotlin/com/adammcneilly/pocketleague/core/datetime/DateTimeFormatter.kt
@@ -1,5 +1,7 @@
 package com.adammcneilly.pocketleague.core.datetime
 
+import kotlinx.datetime.Clock
+import kotlinx.datetime.Instant
 import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.TimeZone
 
@@ -39,7 +41,55 @@ interface DateTimeFormatter {
         return "$minutesStr:$secondsStr"
     }
 
+    /**
+     * Determines if the date represented by the supplied [utcString]
+     * occurred before the current date.
+     */
+    fun isBeforeNow(
+        utcString: String,
+        clock: Clock,
+    ): Boolean {
+        val instant = Instant.parse(utcString)
+
+        return instant < clock.now()
+    }
+
+    /**
+     * Given a [utcString], convert it to a relative timestamp such as 5m ago, or 5d ago.
+     *
+     * @return Only returns a relative timestamp if this date is in the past,
+     * otherwise null.
+     */
+    fun getRelativeTimestamp(
+        utcString: String,
+        clock: Clock,
+    ): String? {
+        if (!isBeforeNow(utcString, clock)) {
+            return null
+        }
+
+        val instant = Instant.parse(utcString)
+
+        val now = clock.now()
+
+        val duration = now.minus(instant)
+
+        return when {
+            duration.inWholeMinutes < MINUTES_PER_HOUR -> {
+                "${duration.inWholeMinutes}m ago"
+            }
+            duration.inWholeHours < HOURS_PER_DAY -> {
+                "${duration.inWholeHours}h ago"
+            }
+            else -> {
+                "${duration.inWholeDays}d ago"
+            }
+        }
+    }
+
     companion object {
-        const val SECONDS_PER_MINUTE = 60
+        private const val MINUTES_PER_HOUR = 60
+        private const val HOURS_PER_DAY = 24
+        private const val SECONDS_PER_MINUTE = 60
     }
 }
